requirements.txt
django==3.0.8
djangorestframework==3.11.0
pyjwt==1.7.1
psycopg2-binary==2.8.5
python-decouple==3.3
flake8==3.8.3
coverage==5.2
coveralls==2.1.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
pytest-django==3.9.0
pytest-cov==2.10.0
mixer==6.1.3